feat: sp3-05-api-handling-not-found

+++ src/controllers/banner.controller.ts
@@ -2,7 +2,7 @@ import { Response } from "express";
 import { IPaginationQuery, IReqUser } from "../utils/interfaces";
 import response from "../utils/response";
 import BannerModel, { bannerDAO, TypeBanner } from "../models/banner.model";
-import { FilterQuery } from "mongoose";
+import { FilterQuery, isValidObjectId } from "mongoose";
 
 export default {
     async create(req: IReqUser, res: Response) {
@@ -56,7 +56,17 @@ export default {
     async findOne(req: IReqUser, res: Response) {
         try {
             const { id } = req.params;
+
+            if (!isValidObjectId(id)) {
+                return response.notFound(res, "failed find one a banner");
+            }
+
             const result = await BannerModel.findById(id);
+
+            if (!result) {
+                return response.notFound(res, "failed find one a banner");
+            }
+
             response.success(res, result, "success find one a banner");
         } catch (error) {
             response.error(res, error, "failed to find one a banner");
@@ -65,6 +75,11 @@ export default {
     async update(req: IReqUser, res: Response) {
         try {
             const { id } = req.params;
+
+            if (!isValidObjectId(id)) {
+                return response.notFound(res, "failed update a banner");
+            }
+
             const result = await BannerModel.findByIdAndUpdate(id, req.body, {
                 new: true,
             });
@@ -76,6 +91,11 @@ export default {
     async remove(req: IReqUser, res: Response) {
         try {
             const { id } = req.params;
+
+            if (!isValidObjectId(id)) {
+                return response.notFound(res, "failed remove a banner");
+            }
+
             const result = await BannerModel.findByIdAndDelete(id, {
                 new: true,
             });
